+++ projects/clr/hipamd/src/hip_hmm.cpp
@@ -379,7 +379,7 @@ hipError_t hipMemRangeGetAttributes(void** data, size_t* data_sizes,
     HIP_RETURN(hipErrorInvalidValue);
   }
 
-  for (int i = 0; i < num_attributes; i++) {
+  for (size_t i = 0; i < num_attributes; i++) {
     if (!data[i]) {
       HIP_RETURN(hipErrorInvalidValue);
     }
