+++ projects/clr/rocclr/device/rocm/rocdevice.hpp
@@ -386,7 +386,7 @@ class Device : public NullDevice {
 
   //! Get the CPU agent that is in a 'index' NUMA node
   const hsa_agent_t getCpuAgent(int index) const {
-    if ((index < 0) || (index >= cpu_agents_.size())) {
+    if ((index < 0) || (static_cast<size_t>(index) >= cpu_agents_.size())) {
       // Return default CPU agent
       return cpu_agent_info_->agent;
     }
